% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{Histogram}
\alias{Histogram}
\title{Histogram function}
\usage{
Histogram(data, x, y = NULL, group)
}
\arguments{
\item{data}{The data frame containing the variables.}

\item{x}{The variable used for the x-axis.}

\item{y}{The variable used for the y-axis.}

\item{fill}{The variable used for fill color.}
}
\value{
A ggplot object displaying a Histogram.
}
\description{
This function creates a Histogram with ggplot2.
}
\examples{
data <- data.frame(
  Peak_Expiratory_Flow = c(250, 300, 280),
  Height = c(160, 170, 165),
  Treatment = c('Placebo', 'Drug', 'Placebo')
)
Histogram(data = data, x = Peak_Expiratory_Flow, y = Height, fill = Treatment)
}
