% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{Scatter plot function}
\usage{
ScatterPlot(data, x, y, fill, label = NULL)
}
\arguments{
\item{data}{The data frame containing the variables.}

\item{x}{The variable used for the x-axis.}

\item{y}{The variable used for the y-axis.}

\item{fill}{The variable used for fill color.}

\item{label}{The variable used for labels. Optional.}
}
\value{
A ggplot object displaying a scatter plot.
}
\description{
This function creates a scatter plot with ggplot2.
}
\examples{
data <- data.frame(
  Peak_Expiratory_Flow = c(250, 300, 280),
  Height = c(160, 170, 165),
  Treatment = c('Placebo', 'Drug', 'Placebo')
)
ScatterPlot(data = data, x = Peak_Expiratory_Flow, y = Height, fill = Treatment)

}
